ACPI: processor: idle: Add debug log for states with invalid entry methods
According to ACPI spec, entry method in LPI sub-package must be a
buffer or an integer.

The driver will disable the state whose the entry method is invalid
by zeroing flags in struct acpi_lpi_state.

The entry method is very key in cpuidle. A debug log is very useful
for developers.

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Reviewed-by: Jonathan Cameron <jonathan.cameron@huawei.com>
[ rjw: Subject and changelog edits, changed "illegal" to "invalid" ]
Link: https://patch.msgid.link/20251125064702.3666149-1-lihuisong@huawei.com
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>